Add a reliability property on "known package list" and use it in app details.
What we wanted to achieve here was that if it is already installed then
show "uninstall" button. Turned out that we cannot really do that because
the app list is not always reliable and not everyone (including ourselves)
cannot get the reliable list of existing apps.

Therefore, by default, the list of existing packages is now marked as
unreliable.

We can conditionally hide "uninstall" button ONLY IF the list of the existing
packages is reliable.

On the other hand, we conditionally show "it is already installed..." text
if the unreliable list of the installed packages contains it, as we do not
tell otherwise.